# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evaluaciones OMR API",
    "description": "Servicio de evaluaciones con reconocimiento óptico de marcas (OMR).\n\n**Flujo principal**\n1. Crear una evaluación (banco de preguntas) → agregar secciones → cargar preguntas con `seccion_id`.\n2. Generar hojas: `POST /generar/{evaluacion_id}` con una selección **template** (reparto por\n   cantidad/porcentaje por sección) o **estructurada** (banco completo o subconjunto ordenado).\n   El `identificador` es opcional: si falta se autogenera (`{short_id}-{n}`) y la reimpresión\n   con el mismo identificador devuelve la misma variante.\n3. Escanear la hoja llena y evaluarla con `POST /evaluar` (califica contra el banco).\n\n**Hoja genérica (sin examen)**\n- `POST /generar/generica` genera solo burbujas (N preguntas × M opciones, 2–5).\n- `POST /leer` devuelve lo llenado (identificador persona + burbujas marcadas) sin contrastar nada.\n\nVer **Redoc** (`/redoc`) para la referencia completa o el esquema OpenAPI en `/openapi.json`.",
    "contact": {
      "name": "Evaluaciones OMR"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/generar/{evaluacion_id}": {
      "post": {
        "tags": [
          "Generación de hojas de examen"
        ],
        "summary": "Generar Hoja",
        "description": "Genera las hojas de un examen a partir del banco.\n\n**Selección**\n- `seleccion.tipo = \"template\"`: reparto por secciones con `cantidad` (exacto) o\n  `porcentaje` (proporción del total, redondeo por mayor resto). El total no puede\n  exceder el banco y siempre es exacto.\n- `seleccion.tipo = \"estructurada\"`: banco completo en su orden, o solo las preguntas\n  de `preguntas_orden` si se indica.\n\n**Identificador (D4)** — opcional. Si falta se genera `{short_id}-{n}`. Llamar de nuevo\ncon el mismo `identificador` devuelve la misma variante (reimpresión estable).\n\n**Respuesta**: PDF de burbujas + hoja de preguntas en base64, y `config_seleccion`\ncon el snapshot de la selección aplicada.",
        "operationId": "generar_hoja_api_v1_generar__evaluacion_id__post",
        "parameters": [
          {
            "name": "evaluacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluacion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerarHojaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerarHojaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerarHojaRequest": {
        "properties": {
          "identificador": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identificador",
            "description": "Identificador de la hoja; si falta se genera '{short_id}-{n}' (D4). Misma (evaluación, identificador) → misma variante (reimpresión estable)."
          },
          "seleccion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeleccionTemplate"
              },
              {
                "$ref": "#/components/schemas/SeleccionEstructurada"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seleccion",
            "description": "Configuración de selección (template | estructurada). Si se omite, se usa el banco completo en orden."
          },
          "cantidad_preguntas": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cantidad Preguntas",
            "description": "Legacy: recorte del banco (equivale a un template sin secciones)"
          },
          "aleatorio": {
            "type": "boolean",
            "title": "Aleatorio",
            "description": "Legacy: selección/orden aleatorios",
            "default": false
          },
          "hojaPreguntas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hojapreguntas",
            "description": "PDF de preguntas (base64) cargado externamente — aún sin uso"
          },
          "hojaRespuestas": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hojarespuestas",
            "description": "PDF de respuestas (base64) cargado externamente — aún sin uso"
          }
        },
        "type": "object",
        "title": "GenerarHojaRequest"
      },
      "GenerarHojaResponse": {
        "properties": {
          "hoja_id": {
            "type": "string",
            "format": "uuid",
            "title": "Hoja Id"
          },
          "identificador": {
            "type": "string",
            "title": "Identificador"
          },
          "qr_data": {
            "type": "object",
            "title": "Qr Data"
          },
          "pdf_base64": {
            "type": "string",
            "title": "Pdf Base64"
          },
          "hoja_preguntas_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hoja Preguntas Base64"
          },
          "cantidad_preguntas": {
            "type": "integer",
            "title": "Cantidad Preguntas",
            "default": 0
          },
          "config_seleccion": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Seleccion"
          }
        },
        "type": "object",
        "required": [
          "hoja_id",
          "identificador",
          "qr_data",
          "pdf_base64"
        ],
        "title": "GenerarHojaResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SeccionQuota": {
        "properties": {
          "seccion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Seccion Id",
            "description": "UUID de la sección del banco"
          },
          "cantidad": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cantidad",
            "description": "Número exacto de preguntas de esta sección"
          },
          "porcentaje": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Porcentaje",
            "description": "Porcentaje del total para esta sección (redondeo por mayor resto)"
          }
        },
        "type": "object",
        "required": [
          "seccion_id"
        ],
        "title": "SeccionQuota"
      },
...
      "SeleccionEstructurada": {
        "properties": {
          "tipo": {
            "type": "string",
            "enum": [
              "estructurada"
            ],
            "const": "estructurada",
            "title": "Tipo",
            "description": "Banco completo en su orden (o subconjunto ordenado)"
          },
          "preguntas_orden": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preguntas Orden",
            "description": "Subconjunto explícito de preguntas en orden; si falta, se usa todo el banco"
          }
        },
        "type": "object",
        "required": [
          "tipo"
        ],
        "title": "SeleccionEstructurada"
      },
      "SeleccionTemplate": {
        "properties": {
          "tipo": {
            "type": "string",
            "enum": [
              "template"
            ],
            "const": "template",
            "title": "Tipo",
            "description": "Selección por secciones (cantidad/porcentaje)"
          },
          "cantidad_preguntas": {
            "type": "integer",
            "title": "Cantidad Preguntas",
            "description": "Total de preguntas a imprimir (debe estar entre 1 y el tamaño del banco)"
          },
          "secciones": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SeccionQuota"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secciones",
            "description": "Reparto por sección: cada item con 'cantidad' o 'porcentaje'"
          },
          "aleatorio": {
            "type": "boolean",
            "title": "Aleatorio",
            "description": "Si true, selección y orden aleatorios (secciones intercaladas)",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "tipo",
          "cantidad_preguntas"
        ],
        "title": "SeleccionTemplate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}